+++ sam.h
@@ -19,8 +19,8 @@ static inline void sam_vdg_hsync(void) {
 void sam_init(void);
 void sam_reset(void);
 void sam_run(int cycles);
-unsigned int sam_read_byte(unsigned int addr);
-void sam_store_byte(unsigned int addr, unsigned int octet);
+uint8_t sam_read_byte(uint16_t addr);
+void sam_store_byte(uint16_t addr, uint8_t octet);
 void sam_nvma_cycles(int cycles);
 void sam_vdg_fsync(void);
 void sam_vdg_bytes(int nbytes, uint8_t *dest);
